Add hooks to be able to collect logs from the SDK
To add a hook, implement the KeenLogSink protocol, call
[[KeenClient sharedClient] addLogSink:myLogSink], then enable logging.
If no log sink is added before logging, a default sink that prints
output to NSLog is added. The level of logging verbosity sent to the
sink can be adjusted by calling
[[KeenClient sharedClient] setLogLevel:<log level>].

Add unit tests. Rename KLL_WARN to KLL_WARNING. Add the log level of
the logged message to the KeenLogSink protocol.