Fix cases where global state wasn't being reset between tests.
KeenClient's sharedClient returns nil when no projectid is set,
which leads to nasty side-effects for test setup. Basically, any
write or read key that was set wasn't getting cleared by setUp
because of this, and neither was currentLocation.

The workaround is to carefully reset state to defaults and set
the projectid to nil last. Changes to sharedClient behavior
proposed by PR #193 will improve this situation because sharedClient
will no longer return nil.