Partition and refactor unit tests, fix several issues.
Make opening the database threadsafe, as that can actually happen from the thread calling into the SDK or the SDK's database operation queue.

Make deletion of event and query data synchronous so as to make subsequent database operations deterministic.

XCTestCases can run in parallel, and the fact that tests are using a full on sqlite store mean that deletion of that store needs to be synchronized, so do that with TestDatabaseRequirement.

Fix a few unit tests that weren't waiting for asynchronous work to complete before finishing.

Remove a few old references to KIOEventStore.