Create default Cartfile and Podfile, move travis build specific scripts to bin/.
This means building projects the normal Carthage or CocoaPods way will work and will pull down the library from github.

Cartfile.resolved and Podfile.lock have ben gitignored so that when a release is made a subsequent update to those files isn't required. The samples should always point to the latest version of the SDK in that branch as well.

Further of note, the samples won't necessarily always build using the provided Cartfile or Podfile since they don't reference the branch version of the SDK. In other words, they'll only be expected to work for the latest release tag in master. Only the travis build scripts will work in a general sense for any branch and any changeset.